Move test command implementation out of app
